change fild and costructor to get double

+++ src/it/unibo/oop/lab05/ex3/ProductImpl.java
@@ -7,7 +7,7 @@ public class ProductImpl implements Product {
 	
 	
 	
-	public ProductImpl(final String name, final int quantity) {
+	public ProductImpl(final String name, final double quantity) {
 		super();
 		this.name = name;
 		this.quantity = quantity;
